requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch
torchvision
stable-baselines3[extra]
gym
box2d-py
hydra-core==1.3.2
pytest==8.3.4
scipy
omegaconf
